+++ Platforms/OpenGL3/OpenGL3Platform/src/MyGUI_OpenGL3VertexBuffer.cpp
@@ -7,7 +7,7 @@
 #include "MyGUI_OpenGL3VertexBuffer.h"
 #include "MyGUI_OpenGL3Diagnostic.h"
 
-#include <GL/glew.h>
+#include <GL/gl.h>
 
 namespace MyGUI
 {